Завершение ЛР6

+++ Graphic_LR_6/Mesh.cpp
@@ -58,7 +58,7 @@ void Mesh::Draw()
 
 	glDrawElements(
 		GL_TRIANGLES, 
-		static_cast<unsigned int>(indices.size()), 
+		indices.size(), 
 		GL_UNSIGNED_INT, 
 		0);
